Change ClassUse to use path when defining caller
In an application, it could happen that two classes are used
during the same run time, and that those two classes' methods
have the same name. ClassUse algorithm to define a class caller
was changed to account for that, but the proposed solution was
not enough, due to the algorithm defining what class was the
caller using '#{path}:#{line_number}'.

This algorithm is problematic when multiple classes are defined
in the same file, or when a class is defined in multiple files.

For that reason, this commit lets the pipeline (rspec) broken,
and a redesign of ClassListing and ClassUse will be made.