Add filter by caller option
FilterUse allowed an array of ClassUse to be filtered using attributes
of the referred object, that is, :method, :name and :path. Filtering
a class use by its :caller_class is also an somethin that should be
allowed.

This is what this commit does: allow a class use to be filtered by
its direct or indirect callers.