Add top of the stack to attribute to class use
Running accessed_classes from ClassLister could show a lot of classes,
but it might be desirable that some of the results don't appear multiple times, as
they take part in other stacks.

To allow that filtering, ClassUse received the attribute 'top_of_stack',
and ClassLister settles whether a use is top of a stack or not.