Parse class name depending on event
Classes names depend on the event that they represent: for a method
call, the name comes from TracePoint#defined_class. For a block call,
the name must be inferred from self. The class from block call is also a
singleton, meaning that it's name has the format <CLASS:(...)>.
Therefore, :b_call names must be parsed to be in the same format as
:call names.