Change top and bottom of stack check methods
ClassUse#matches_top_of_stack depended on an attribute from ClassUse to
work. With the addition of callees, if a ClassUse is a top_of_stack can be
inferred from the number of callees that a class has.

In the same way that the top_of_stack can be inferred from callees,
bottom_of_stack can be inferred from the callers that a class has. This
commit adds methods to check both of those properties.