Format classes uses to express their relationship
To get to the main goal of this Gem, which is showing the runtime
relationship of classes, an array of ClassesUses must be parsed to
express the relationship between the classes uses.

This is what the ClassesRelationshipParser does: it generates a hash
that indicates which class is caller, and which class is callee.