Generate a report for classes uses relationship
Using a template with D3.js, trace2 is able to generate a report of the
relationship between classes, which is the main goal of this Gem.

To do that, ReportGenerator must receive an output path, an array of
classes uses and a class that processes the relationship between classes
uses.