Add C extensions to the project
Due to the performance hit that this gem causes when ran with a common
gem, like sinatra or rspec, it wouldn't be viable to keep going with the
project, unless the performance is improved.

Also, the issue in which classes override '.to_s' and there is no way
around could turn out to be a huge issue, as it caused some experimental
runners to break.

To fix both issues, C extensions where introduced. This first commit
introduces C code that shouldn't be used.