Add class_use struct to event_processor
Both class_list and class_stack were more than mere wrappers to the use
of a class, when both of them should only reference the structure that
contained data about the use of classes. The class_use is the struct
that both data structures wrap.

It's also important to notice that the ruby application isn't working as
expected. That will only happen when event_processor.c is entirely
refactored.