Add ruby functions to event_processor.c
The extension has functions that should be used by Ruby only, like the
initialize function. The functions that were added to the extension
should conform to the interface that Trace2::EventProcessor has.

The extension code should also be tested by rspec tests. Furthers tests
will be added in the future to guarantee that the extension functions
as expected.