Add validate_name to query_use
query_use.c will need to have all validations that QueryUse class has.
The first of the validations that was added is validate_name, which
validates if a class use name is equal to a name inside a list of
names.

With the creation of the function and it's tests, many files had to be
added, and that was also done in this commit.