Add further simple validation methods to query_use
Beside name, class_use has other field that could be easily validated.
Those methods validate if method, path and lineno are equal to any of
the values passed in an array.

A feature on the roadmap is adding RegEx or check if a string is
contained by another to check validate a value.