Add count_occurrences to query_use.c
Count occurrences will be a helper function to build a filter, the main
goal of query_use. This function is quite simple: it counts the number
of times a word appears in an array of words, given a start and an end
index. This will help to allocate the memory used by the filter's
structs.