Create filter_builder on query_use.c
The filter_builder function creates a filter given an array of strings,
which will be the input given by Ruby to the extension. The filter
serves to partially substitute the functionality that the QueryUse class
offers to native Ruby code.

The functionalities that it doesn't cover is to allow the filtering of a
class_use by it's caller or callees, and using a regex to validate the
value of the class_use.

filter_builder did not need some functions that were defined in
query_use, thus making them useless. They need to be removed.