Initialize ClassLister with a hash
ClassLister can receive different event processors, filters and filter
parsers, and the list of input arguments might grow in the future. To
improve the class lister extensibility, ClassLister was changed to be
initialized with a hash.

The defaults were defined to not break tests that already exist and
prioritize Ruby classes.