Enable filters on EventProcessorC
The filtering functionality that the native Ruby code has was partially
enabled on the extension code. To use it, the FilterParser should
be applied on a valid filter and passed to the EventProcessorC class.

The things that are still not offered by the extension future are filtering
by caller and callee and using regex on the values used by the filters.