Fix frequent segfaults due to query_use.c
The lest test added to event_processor_c_spec generated random
segfaults, which were hard to detect. To make debugging that kind of
issue easier, a script to run Valgrind was added.

After running Valgrind, the bug was found and fixed. It was due to a
type name being used as a variable inside the function that allocates
memory for validations.

The scripts that are used to help on the development were also changed
to split functionality over more scripts.