Add bottom and top of stack validations
To validate if a class is the top of stack, that is, if it does not call
any other class, and to validate if a class is bottom of stack, that is,
if a class isn't called by any other class, the botom and top of stack
validations were added to the extension. Both of those validations were
already available on the ClassUse class.

The FilterParser class had to be adapted to correctly parse the new
validations.