Create caller_class filter on the extension
The QueryUse class supported the caller_class filter, which validates a
filter against all direct and indirect callers of a class. That filter
was added to the extension. In order to do that, Ruby and C tests had to
be written.

During the write of the extension's tests, a bug was found in which the
callers of a class were wrongfully registered. That happened due to the
wrong events being used on the ClassLister class. To fix that, the
method "#event" was added to all EventProcessor classes.