Add regex support to the extension's filter
The native Ruby code already gave support to it's filters, while the
extension code didn't have it. To make the two of them equal, the
extension started to support regexes. In order to do that, the Onigmo
header, that is, the header of Ruby's native regex support, had to be
added to the extension, and a C file named regex.c were also added. The
isolation of regex functions and types on that file and it's header
enable a more decoupled code.

The equivalents to regex.c that allowed the extension tests to continue
run without major changes to the tests were also created on the ext/test
folder.