Create Trace2::Summarizer extension class
Trace2::Summarizer will be used to summarize the results that
ClassLister gets. It is currently incomplete and useless, but adding
the boilerplate code required for it to work and be properly tested
is enough to make a commit. This new class  will work only for the
extension and shows the intent of prioritizing extension code over Ruby
code.