Create ext function to summarize classes uses
To generate the files that show the relationships between classes, it's
necessary to reduce classes uses lists. The function that got added to
summarizer.c, reduce_uses_list, reduces a class use list to a
summarized_list, in which each element has a class name and a list of
methods that got called.