Change summarized_list generation output
The reduce_uses_list generated an output that didn't contain all data
required to analyze the classes uses of a process, and in order to
achieve a format that could be easily used, another function would be
required. To avoid that, the summarized_list struct was changed to contain
both caller and callee name.

That method also got renamed to reduce_callees_list, due to it's primary
responsibility being reducing a list of callees.