Create summarize function on the extension
The reduce_callees_list function that was used by the extension
created a new data structure, but that data structure didn't have
all the data required to create the report, and the data also had
a format that was easy to use.

To address those issues, the summarize function was created on
the extension along with data structures that have all the
data that will be used to generate a report.