Add reducers to summarizer extension
To make it easier to generate the summary of the relationship between
classes, two functions were added to the summarizer.c extension: the
reduce_class_methods_list and the reduce_summarized_uses. While the
first method reduces a list of classes methods, the second method
depends on the first method and reduces a list of summarized uses.

reduce_class_methods_list gets classes methods that have the same name,
concatenate their methods' lists and get only the unique methods, while
the reduce_summarized_uses gets summaries with the same class use name
and reduces it's methods, callers and callees list.