Implement an extension hash table
To be able to make a graph of the relationship between classes, it's
important to only create nodes of elements that are not duplicate. A
fast way to get a list of unique elements of a list is by using a hash table.
With a hash table that uses a good hashing function and is big enough,
the memory cost of using a hash table will be paid by it's fast
execution time, which will be better than bruteforcing for repeated
elements or somehow sorting the list of elements that will be used as
nodes.