Create Trace2::GraphGenerator
Trace2::GraphGenerator is the class that will generate a graph file that
will somehow be rendered. To mantain the goal of a gem that is fast
enough to minimize performance issues, the class will exist on the
extension.

This commit only added the boilerplate code required to use an extension
class.