Create graph_strings function to graph_generator
Given a class_use, the relationship that it has to other classes should
be expressed in a series of strings, that end with a NULL pointer.
That is what the graph_strings method returns for a given class use.
That method will be used to generate the graph representation of
relationships between caller and callees.