Create build_graphs_array function to extension
To get all unique caller-callee relationships that occur during the
execution of a program, the build_graphs_array function was added to the
extension. That function creates an array of string with format
"caller -> callee" by using a classes_list struct, that is, a
linked-list of class_use. To generate the uniques list, the hash_table
created on hash_table.c was used.

Hash table had it's hashing function changed to retrn positive integers,
making it easier to use the hash result on a the hash_table.