Create write_graph_file on extension
After creating the array of strings that have unique caller-callee
relationships between classes, the results should be written somewhere.
The responsibility of writing these relationships on a file is done by
the write_graph_file function that exists on the graph_generator file.

The function relies on C's stdio.h to avoid incurring on Ruby's
overhead and help keep the code simple.