Add graph creation to Trace2::GraphGenerator
Trace2::GraphGenerator#run was created to allow the generation of a
directed graph file that describes the caller-callee relationships
encountered during runtime. The graph that is generated by that
method follows the DOT language, a graph description language.

The extension code was changed to allow the generation of valid directed
graphs.