Remove old template and report classes
ReportGenerator, a class that should generate a report of classes uses,
and both RelationsipParser::D3 and RelationshipParser::Mermaid, which
parsed the caller-callee relationships to specific formats, were
scrapped for many reasons: their performance was slow, the exhibition
of their results could be really slow or even impossible on a browser
and the format that was being used added a dependency to Javascript
libraries.

The responsibilities that those classes held went to GraphGenerator,
making both generation and exhibition of graphs faster. The dependency
to external libraries remain, but there is the benefit of not depending
on a browser to render the report.