Create Trace2::Runner
The Trace2::Runner class is used to run another runner and generate a
report of the caller-callee relationships that showed up during the
second runner's callback. For example, the following usage:
```
trace2 rspec spec/
```

Runs rspec on the spec/ folder, listing the classes that were used
during rspec's execution. It also generates a report that describes
the existing relationships. The current version of the gem is only able
to generate a report that is a text file that uses the DOT language.
Future versions of the gem will have use OptionsParser to provide more
functionalities.

The 'at_exit' method is used to prevent that a runner exits the process
and no report is generate.

spec/spec_helper.rb and .gitignore were updated to improve testing.