Add runner support to Ruby files
Besides supporting executables, trace2 should be able to generate a
report for valid Ruby scripts. The Trace2::Runner class was changed
to be able to run those files. A more detailed was also added to notify
the user that a file doesn't have valid syntax.

The bin/trace2 file was also changed to not user parameters, letting
ARGV be the only object used by the runner.