Create OptionParser
OptionParser is used along Trace2::Runner to parse the options
passed to the command line executable. The Trace2::OptionParser class
has, for now, simple options: version, help, output path and filter
path.

The class returns a hash with all the options that will be used used by
the Runner.

A version file and an script to update the gem's version were added to
make version update easier.