Create Options and OptionParser
Trace2::OptionParser was split into Trace2::OptionParser and
Trace2::Options. While OptionParser is based on Ruby's standard
OptionParser and has the extra functionality of spliting the arguments
that it receives into trace2 arguments and another executable's
arguments, the Option class defines the options and parses the arguments
that it receives.