Integrate OptionParser with Runner
OptionParser was not included on Runner, therefore no option was passed
to it. Including the OptionParser on Runner.run method made it possible
to use the inputed options.

During the inclusion of OptionParser, the ExecutableRunner class was
extracted from Runner, and it is responsible for running the executable
that is passed to Runner. With that change, classes have less
responsibilities and tests are simpler.

The project also had rubocop-rspec added to it and the standard cops
will be used. That decision was taken to enforce style on the tests,
improving code maintainability and readability. For the time being,
rubocop still has offenses, which will be fixed before trace2 1.0.0.

Lastly, trace2.rb file was updated to include the files that were
created on the last few commits, fixing the gem.