Create ClassListerBuilder
The creation of a ClassLister involves an event processor (either
EventProcessor or EventProcessorC) and a filter parser (either
FilterParser on nil). As one type of event processor is always used with
the same type of filter parser, a builder is used to make it easier to
create a class lister.

To be able to use that new structure, FilterParser interface had
to be changed and it's tests were refactored to use the new interface.
Tests that relied on the old interface of FilterParser remain broken and
will be refactored on the next commits.

spec/spec_helper.rb was changed to require lib/trace2.rb instead of each
individual files, thus being more similar to how the Gem works.

spec/trace2/arguments_parser_spec.rb and lib/trace2/arguments_parser.rb were
removed because they will no longer be used.