Fix tests after ClassLister's new builder class
The ClassListerBuilder class improved maintainability and
responsibilities segregations on the application, but it also changed
how a ClassLister is instantiated, thus breaking tests that relied on
the old instantiation format. Due to that, the EventProcessor and
EventProcessorC had to be fixed. The tests of those classes were also
split into integration and regular tests, improving how tests are
organized.

ClassListerBuilder was also update to receive a filter instead of a
filter_path, making the ClassLister easier to test.

Lastly, graph_generator_spec.rb test had to be fixed for a different reason:
the graph_generator.c extension function was changed to generate
classes' names inside quotes and the Ruby tests were not update to
consider that. That fix was done on this commit.