Integrate ClassListerBuilder with Runner
The last steps required for the integration of ClassListerBuilder with
Runner are implementing a new option on the Option class that receives
which type of EventProcessor will be used on the ClassLister, changing
how a ClassLister is built on the Runner class and changing how the
Runner class fetches the filter that will be used by an EventProcessor.
Those steps were performed on this commit.

Besides those changes, the tests of spec/trace2/options_spec.rb and
spec/trace2/runner_spec.rb had their rubocop offenses fixed. The Options
class still has tests that don't behave as expected when the
expectations are separated into two different 'it' blocks. Those issues
were marked as TODOS, as the investigation of what causes it demands
time and brings little to no value.