Create a class to wrap dot's graph rendering
Rendering a graph that summarizes the relationship between classes is
performed by dot, which is part of graphviz. To hide this complexity
from the user, the DotWrapper class was created. It checks if the system
has dot installed, and if it has, runs the dot executable to render a
graph.

That class was integrated with the Runner class, which should receive an
option (automatic_render) to run the wrapper or not. The Options class
was updated to accept the --manual (or -m) option, that defines if the
wrapper should be run. The wrapper is opt-out, as generating a visual
representation of the graphs brings great value to the user, but it can
increase the total runtime of the gem.