Create GraphGenerator for Ruby
This gem used to have a GraphGenerator that was implemented on the
using native code, and that class only worked when used with the
EventProcessorC. That class got renamed to GraphGeneratorC, and a class
the works with EventProcessor was implemented. That new class is
GraphGenerator.

The intention with those changes is making it possible for the user to
make Ruby extensions to the functionalities offered by this Gem, even
when the Ruby code is way slower than the extension code. Future commits
will create ways to guarantee that the executable always makes the right
combinations, that is, either GraphGenerator with EventProcessor, or
GraphGeneratorC with EventProcessorC.