Change ClassListerBuilder to ReportingToolsFactory
While the ClassListerBuilder satisfied it's initial use, the use of a
EventProcessor is always linked to a GraphGenerator. For instance,
GraphGeneratorC should always be used with a ClassLister that was
initialized with EventProcessorC. Due to that, an abstract factory that
builds and returns a GraphGenerator with a ClassLister that was
initialized with the right type of EventProcessor would improve the
design of the gem. For that reason, the ClassListerBuilder class was
renamed to ReportingToolsFactory, and it now returns a ClassLister and a
GraphGenerator.

Runner class and it's tests had to be changed to use the new format that
ReportingToolsFactory uses. It was also refactored to remove methods
that supplied little functionality. Tests that relied on
ClassListerBuilder were edited to use the new interface that
ReportingToolsFactory supplies.