Enable all types of GraphGenerator on Runner
After implementing ruby's GraphGenerator class, it was necessary to
enable it's usage by the Runner class. To do that, GraphGenerator's
interface was changed to receive two parameters: an output path and a
ClassLister instead of an output path and an array of classes uses. With
that change, it's up to the GraphGenerator classes how to extract the
classes uses from the ClassLister.

To make GraphGenerator and GraphGeneratorC interfaces equal, an extra
parameter was added to GraphGeneratorC#run. That parameter is not used
by the method, but that should be changed in the future, as a parameter
that is not used is bad design. After making the interfaces equal, it
is possible to call the two classes by passing the same parameters on
the Runner class.

Runner class had to be changed to use GraphGenerators' new interfaces.
GraphGenerator, GraphGeneratorC and Runner tests also had to be changed
to adapt to the new interfaces.