Fix Options help and version options
Both help and version options crashed trace2's runner when no
executable was passed. To fix that, the raising of errors due to the
lack of executable was moved to the Options class, and it's executed
after OptionParser parses the options. That change of order guarantees
that the Runner will exit before verifying if there is no executable for
the help and version options.

Besides that change, OptionParser was adapted to accept an array of text
as it's description. That change allows descriptions to have line
breaks, thus improving the options readability.

Finally, :event_processor_type option was changed to :tools_type, making
the option's meaning more intuitive.