Fix most rubocop warnings
After adding rubocop-rspec to the project, some of the tests had cops
that were not satisfied. This commit fixes most of those issues by
refactoring the tests to use the DRY style, which was already used by
the remainder of the project. EventProcessor will still be refactored to
fix the cops that are stiill being broken.