Refactor EventProcessor
The remaining error that rubocop had were due to EventProcessor and that
class had some unneeded complexity. Due to that, it got refactored to
remove methods that were too shallow, or methods that were too complex.
Besides that, all dependencies that EventProcessor uses were moved to
it's constructor, thus making it slightly more decoupled from other
classes.

README got updated to include the generation of an YAML that could be
used as a filter. This gem's version also got a bump to 0.8.0, as few
things are missing to it's launch.