Add output format option
The format option, which determines the format that a .dot file should
be rendered into, was missing on the Options class and thus added to the
project on this commit. Besides that, that class went through some
minor improvements: some default options were added, and the
descriptions for those options were also changed to inform what their
default value.

To be able to use those new default values and the new options, the
Runner class and it's tests had to be adapted.