Add travis-ci support
Travis-CI was added to the project to allow pipelines to be
run automatically, thus helping improve code maintainability and quality.
For the time being, only Ruby versions that did not reach EOL will be
verified, and code quality will be always verified against Ruby's most
recent version.

For now, extension tests will not be run, as they provoke errors and are
unpredictable on the CI environment. In the future, those tests will be
either discarded, as Ruby tests will cover all required functionalities,
or fixed, so they work as expected on any environment.

rubocop had to be changed to ignore /vendor folder. That way,
false-negatives on travis-ci will be avoided.

OptionParser had to have it's merge method changed, as in Ruby 2.5, the
Hash#merge method didn't accept more than one hash.