Fix trace2.so require
To be able to use multiple versions of Ruby with trace2, the way that
the trace2.so file is required was changed to consider Ruby's version.
It follows the conventions presented by rake-compiler.

For the time being, only the x86_64-linux platform is supported, as the
efforts with rake-compiler and rake-compiler-dock were not successful.
If this gem finds more usage, support for other platforms might be
offered.

The trace2_c.so that was generated by compiling the extension code was
changed to trace2.so, as the .so file having the same name as the gem
makes it more intuitive.

Rakefile and .gemspec were changed to try to use rake-compiler's
cross-compiling resource, that was ditched.